Add files via upload
I now load the new_main_window ui file. No longer need to use the setup ui function in new_main_window. Merged with Nawaf's latest changes to show cards based on category and to search cards in the search bar. The forgot password functionality has yet to be implemented. There is a bug when we add a new card of a different category when we are in a specific category. Currently there is no way to tell which category we are currently in on the user interface. The ui appearance has been updated and all buttons on the interface have been implemented.